Add Game Analytics Initializer
This removes the need for putting the Game Analytics prefab in your startup scene, instead spawning the initializer from any scene you start in, allowing for a better editor-logging workflow and more freedom in the way you set up your game.

This solution offers a couple of benefits:
1. Game Analytics becomes more plug-and-play.
2. Extension is easily achieved by providing your own GameAnalytics prefab at path Resources/Prefabs/GameAnalytics. Custom resources take priority over packaged resources.
3. Start Game Analytics from any scene.

The downside of this is that Game Analytics has quite a few debug logs, which may be undesirable. My suggestion would be to update GA in order to make editor debugging optional.